refactor: select 코드 분리
